web-animations/animation-player/animation-player-paused.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer paused attribute tests</title>
<meta name="assert" content="A boolean value that is true if the player is currently paused.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-paused">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = newAnimationPlayer(createDiv(this));

    player.currentTime = ANIMATION_END_TIME - 1;
    assert_false(player.paused, 'The value of AnimationPlayer.paused attribute should be ' +
        'false if player is not paused');

    player.currentTime = ANIMATION_END_TIME + 1;
    assert_false(player.paused, 'The value of AnimationPlayer.paused attribute should be ' +
        'false if player is not paused but seek time is greater than content end');
}, 'The value of AnimationPlayer.paused attribute is false if player is not paused');

test(function() {
    var player = newAnimationPlayer(createDiv(this));

    player.pause();
    assert_true(player.paused, 'AnimationPlayer should be paused');

    player.play();
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    player.pause();
    assert_true(player.paused, 'AnimationPlayer should be paused again');
}, 'The value of AnimationPlayer.paused attribute is true if player is paused');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = ANIMATION_END_TIME - 10;
    setTimeout(function() {
        t.step(function () {
            assert_false(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be false');
            player.pause();
            assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be true');
            t.done();
        });
    }, 50);
}, 'Test AnimationPlayer.paused attribute if player is finished playback rate > 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = 10;
    player.playbackrate = -1;
    setTimeout(function() {
        t.step(function () {
            assert_false(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be false');
            player.pause();
            assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
                'should be true');
            t.done();
        });
    }, 50);
}, 'Test AnimationPlayer.paused attribute if player is finished and playback rate < 0');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = ANIMATION_END_TIME - 100;
    player.pause();
    animation.timing.playbackRate = 2;
    assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
        'should be true');
}, 'Test AnimationPlayer.paused attribute if animation length is changed (shortened)');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.currentTime = ANIMATION_END_TIME + 100;
    player.pause();
    animation.timing.playbackRate = 0.5;
    assert_true(player.paused, 'The value of AnimationPlayer.paused attribute ' +
        'should be true');
}, 'Test AnimationPlayer.paused attribute if animation length is changed (extended)');
</script>
</body>
